refactor: slots refactor
