+++ src/slots.ts
@@ -1,0 +1,6 @@
+import { SlotOperation } from '@openedx/frontend-base';
+
+const slots: SlotOperation[] = [
+];
+
+export default slots;
