+++ src/slots/FooterSlot/README.md
