+++ src/slots/FooterSlot/images/custom_footer.png
