+++ src/slots/FooterSlot/images/default_footer.png
