+++ src/slots/IdVerificationPageSlot/images/custom_id-verification-page.png
