+++ src/slots/IdVerificationPageSlot/images/default_id-verification-page.png
