+++ src/slots/README.md
@@ -1,4 +1,4 @@
 # `frontend-app-account` Plugin Slots
 
 * [`org.openedx.frontend.layout.footer.v1`](./FooterSlot/)
-* [`org.openedx.frontend.account.id_verification_page.v1`](./IdVerificationPageSlot/)
+* [`org.openedx.frontend.slot.account.idVerificationPage.v1`](./IdVerificationPageSlot/)
